Initial refactor for ConfigDocumentFactory
