+++ sconfig/jvm-native/src/main/scala/org/ekrich/config/parser/ConfigDocumentFactoryJvmNative.scala
@@ -1,0 +1,6 @@
+package org.ekrich.config.parser
+
+/**
+ * [[ConfigFactoryDocument]] methods common to JVM and Native
+ */
+abstract class ConfigDocumentFactoryJvmNative extends ConfigDocumentFactoryShared {}
