+++ sconfig/jvm/src/main/scala/org/ekrich/config/parser/PlatformConfigDocumentFactory.scala
@@ -1,0 +1,6 @@
+package org.ekrich.config.parser
+
+/**
+ * [[ConfigFactoryDocument]] methods for Scala JVM platform
+ */
+abstract class PlatformConfigDocumentFactory extends ConfigDocumentFactoryJvmNative {}
