+++ sconfig/native/src/main/scala/org/ekrich/config/PlatformConfigFactory.scala
